Beta Version of the Train Models
This version is the first beta version of the training pipeline. The main changes tackle saving models and forecasting from pre-trained models using the MinIO server as storage option.

A config file has been added to allow setting models within the minIO server.

Outside of this the parameters of the pipeline have been simplified.